Search driver genes

This function depends on a chromosome files folder, hg19 or hg38, which includes the txt files of each chromosome. The hg19 folder is able to download at chr_files_hg19. This function first preprocesses input data via the preprocessing function. Then the BMR function is used to calculate the background mutation rate. In order to search driver genes, there are 6 hypothesis test methods to sellected from according to the parameter p_class, i.e. binomial distribution test, represents beta binomial distribution test, Fisher combined P-value test, likelihood ratio test, convolution test and 2D-projection method, which are all encapsulated in the sigGenes function. The output files contatins txt files including the preprocessed mutation and coverage data, mutation categories, and significant driver genes. There are several plots output as pdf files as well.

DriverGenes(Mutation = NULL, Coverage = NULL, Covariate = NULL,
  MutationDict = NULL, chr_files_directory = "chr_files_hg19",
  categ_flag = NaN, bmr = 1.2e-06, p_class = "allTest",
  sigThreshold = 0.05, file_stem = "text")

Arguments

Mutation
Mutation maf data, default to be NULL.

Coverage
Coverage raw data, default to be NULL.

Covariate
Covariate data which includes expression level, replication time and chromatin compartment defaultly, and could be altered by users, default to be NULL.

MutationDict
Mapping Variant_Classification to mutation effect in mutation data, default to be NULL.

chr_files_directory
Chromosome files directory, hg19 or hg38.

categ_flag
Mutation category number, should be either NaN or numeric, default to be NULL.

bmr
The default background mutation rate is 1.2e-6, and the value alters when function ends.

p_class
Hypothesis test methods. "binomial" represents binomial distribution test, which is default; "betabinomial" represents beta binomial distribution test; "fisher" represents Fisher combined P-value test; "LRT" represents likelihood ratio test; "CT" represents convolution test; "projection" represents 2D-projection method; "allTest" represents the mutual results of all methods.

sigThreshold
The threshhold of q-value to judge if the gene is significant.

file_stem
Filestem